Move some functions to utils.py
